Enhance scoring system for coding tests with decimal precision
- Updated models to use decimal types for scores and marks across various entities, ensuring accurate representation of scoring.
- Introduced ScoringOptions class for configurable default marks based on problem difficulty.
- Modified CodingTestService to utilize ScoringOptions for default marks assignment during test creation.
- Added OutputComparer for precise output comparison in judge evaluations.
- Refactored score calculation logic in PerformanceService and PracticeTestService to handle decimal scores.

+++ LeetCodeCompiler.API/Models/CodingTestAttempt.cs
@@ -28,9 +28,9 @@ namespace LeetCodeCompiler.API.Models
         [Required]
         public string Status { get; set; } = "InProgress"; // InProgress, Completed, Submitted, Abandoned
 
-        public int TotalScore { get; set; } = 0;
+        public decimal TotalScore { get; set; } = 0;
 
-        public int MaxScore { get; set; } = 0;
+        public decimal MaxScore { get; set; } = 0;
 
         public double Percentage { get; set; } = 0.0;
